requirements.txt
google-api-python-client==2.149.0
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.3
pyarrow==17.0.0
scikit-learn==1.5.2
lightgbm==4.5.0
joblib==1.4.2
tqdm==4.66.5

